+++ packages/amazon-cognito-identity-js/index.d.ts
@@ -70,16 +70,12 @@ declare module 'amazon-cognito-identity-js' {
 	export class CognitoUser {
 		constructor(data: ICognitoUserData);
 
-		public setSignInUserSession(
-			signInUserSession: CognitoUserSession
-		): void;
+		public setSignInUserSession(signInUserSession: CognitoUserSession): void;
 		public getSignInUserSession(): CognitoUserSession | null;
 		public getUsername(): string;
 
 		public getAuthenticationFlowType(): string;
-		public setAuthenticationFlowType(
-			authenticationFlowType: string
-		): string;
+		public setAuthenticationFlowType(authenticationFlowType: string): string;
 		public getCachedDeviceKeyAndPassword(): void;
 
 		public getSession(
@@ -164,7 +160,10 @@ declare module 'amazon-cognito-identity-js' {
 		public sendMFACode(
 			confirmationCode: string,
 			callbacks: {
-				onSuccess: (session: CognitoUserSession, userConfirmationNecessary?: boolean) => void;
+				onSuccess: (
+					session: CognitoUserSession,
+					userConfirmationNecessary?: boolean
+				) => void;
 				onFailure: (err: any) => void;
 			},
 			mfaType?: string,
@@ -184,15 +183,9 @@ declare module 'amazon-cognito-identity-js' {
 			callbacks: {
 				onSuccess: (session: CognitoUserSession) => void;
 				onFailure: (err: any) => void;
-				mfaRequired?: (
-					challengeName: any,
-					challengeParameters: any
-				) => void;
+				mfaRequired?: (challengeName: any, challengeParameters: any) => void;
 				customChallenge?: (challengeParameters: any) => void;
-				mfaSetup?: (
-					challengeName: any,
-					challengeParameters: any
-				) => void;
+				mfaSetup?: (challengeName: any, challengeParameters: any) => void;
 			},
 			clientMetadata?: ClientMetadata
 		): void;
@@ -258,14 +251,8 @@ declare module 'amazon-cognito-identity-js' {
 			callbacks: {
 				onSuccess: (session: CognitoUserSession) => void;
 				onFailure: (err: any) => void;
-				mfaRequired?: (
-					challengeName: any,
-					challengeParameters: any
-				) => void;
-				totpRequired?: (
-					challengeName: any,
-					challengeParameters: any
-				) => void;
+				mfaRequired?: (challengeName: any, challengeParameters: any) => void;
+				totpRequired?: (challengeName: any, challengeParameters: any) => void;
 			}
 		): void;
 	}
@@ -407,4 +394,10 @@ declare module 'amazon-cognito-identity-js' {
 	}
 
 	export const appendToCognitoUserAgent: (content: string) => void;
+
+	export class WordArray {
+		constructor(words?: string[], sigBytes?: number);
+		random(nBytes: number): WordArray;
+		toString(): string;
+	}
 }