+++ packages/amplify-ui-components/src/components/amplify-checkbox/amplify-checkbox.stories.tsx
@@ -4,21 +4,21 @@ import { knobs } from '../../common/testing';
 const checkboxStories = storiesOf('amplify-checkbox', module);
 
 checkboxStories.add('with label', () => {
-  const label = knobs.labelKnob('Boise');
+	const label = knobs.labelKnob('Boise');
 
-  return `<amplify-checkbox label=${label} name="boise" field-id="boise" value="boise"></amplify-checkbox>`;
+	return `<amplify-checkbox label=${label} name="boise" field-id="boise" value="boise"></amplify-checkbox>`;
 });
 
 checkboxStories.add('toggle check box', () => {
-  const label = knobs.labelKnob('Boise');
-  const toggle = knobs.toggleKnob(false);
+	const label = knobs.labelKnob('Boise');
+	const toggle = knobs.toggleKnob(false);
 
-  return `<amplify-checkbox label=${label} name="boise" field-id="boise" value="boise" checked=${toggle}></amplify-checkbox>`;
+	return `<amplify-checkbox label=${label} name="boise" field-id="boise" value="boise" checked=${toggle}></amplify-checkbox>`;
 });
 
 checkboxStories.add('disabled', () => {
-  const label = knobs.labelKnob('Portland');
-  const disabled = knobs.disabledKnob(true);
+	const label = knobs.labelKnob('Portland');
+	const disabled = knobs.disabledKnob(true);
 
-  return `<amplify-checkbox label=${label} name="portland" field-id="portland" value="portland" disabled=${disabled}></amplify-checkbox>`;
+	return `<amplify-checkbox label=${label} name="portland" field-id="portland" value="portland" disabled=${disabled}></amplify-checkbox>`;
 });