+++ packages/amplify-ui-components/src/components/amplify-oauth-button/amplify-oauth-button.tsx
@@ -6,23 +6,26 @@ import { FederatedConfig } from '../../common/types/auth-types';
 import { Translations } from '../../common/Translations';
 
 @Component({
-  tag: 'amplify-oauth-button',
-  shadow: true,
+	tag: 'amplify-oauth-button',
+	shadow: true,
 })
 export class AmplifyOAuthButton {
-  /** Federated credentials & configuration. */
-  @Prop() config: FederatedConfig['oauthConfig'] = {};
+	/** Federated credentials & configuration. */
+	@Prop() config: FederatedConfig['oauthConfig'] = {};
 
-  private signInWithOAuth(event) {
-    event.preventDefault();
-    Auth.federatedSignIn();
-  }
+	private signInWithOAuth(event) {
+		event.preventDefault();
+		Auth.federatedSignIn();
+	}
 
-  render() {
-    return (
-      <amplify-sign-in-button onClick={event => this.signInWithOAuth(event)} provider="oauth">
-        {this.config.label || I18n.get(Translations.SIGN_IN_WITH_AWS)}
-      </amplify-sign-in-button>
-    );
-  }
+	render() {
+		return (
+			<amplify-sign-in-button
+				onClick={event => this.signInWithOAuth(event)}
+				provider="oauth"
+			>
+				{this.config.label || I18n.get(Translations.SIGN_IN_WITH_AWS)}
+			</amplify-sign-in-button>
+		);
+	}
 }