+++ packages/amplify-ui-components/src/components/amplify-select/amplify-select.stories.tsx
@@ -3,5 +3,5 @@ import { storiesOf } from '@storybook/html';
 const selectStories = storiesOf('amplify-select', module);
 
 selectStories.add('default with empty select', () => {
-  return `<amplify-select></amplify-select>`;
+	return `<amplify-select></amplify-select>`;
 });