+++ packages/datastore/src/storage/adapter/AsyncStorageAdapter.ts
@@ -97,7 +97,7 @@ export class AsyncStorageAdapter implements Adapter {
 	async save<T extends PersistentModel>(
 		model: T,
 		condition?: ModelPredicate<T>
-	): Promise<[T, OpType.INSERT | OpType.UPDATE, T?][]> {
+	): Promise<[T, OpType.INSERT | OpType.UPDATE][]> {
 		const modelConstructor = Object.getPrototypeOf(model)
 			.constructor as PersistentModelConstructor<T>;
 		const storeName = this.getStorenameForModel(modelConstructor);
@@ -133,7 +133,7 @@ export class AsyncStorageAdapter implements Adapter {
 			}
 		}
 
-		const result: [T, OpType.INSERT | OpType.UPDATE, T?][] = [];
+		const result: [T, OpType.INSERT | OpType.UPDATE][] = [];
 
 		for await (const resItem of connectionStoreNames) {
 			const { storeName, item, instance } = resItem;
@@ -145,11 +145,7 @@ export class AsyncStorageAdapter implements Adapter {
 			if (id === model.id || opType === OpType.INSERT) {
 				await this.db.save(item, storeName);
 
-				if (opType === OpType.UPDATE) {
-					result.push([instance, opType, fromDB]);
-				} else {
-					result.push([instance, opType]);
-				}
+				result.push([instance, opType]);
 			}
 		}
 
@@ -490,7 +486,7 @@ export class AsyncStorageAdapter implements Adapter {
 		deleteQueue: { storeName: string; items: T[] }[]
 	): Promise<void> {
 		for await (const rel of relations) {
-			const { relationType, modelName } = rel;
+			const { relationType, modelName, targetName } = rel;
 			const storeName = this.getStorename(nameSpace, modelName);
 
 			const index: string =
@@ -510,9 +506,14 @@ export class AsyncStorageAdapter implements Adapter {
 			switch (relationType) {
 				case 'HAS_ONE':
 					for await (const model of models) {
+						const hasOneIndex = index || 'byId';
+
+						const hasOneCustomField = targetName in model;
+						const value = hasOneCustomField ? model[targetName] : model.id;
+
 						const allRecords = await this.db.getAll(storeName);
 						const recordToDelete = allRecords.filter(
-							childItem => childItem[index] === model.id
+							childItem => childItem[hasOneIndex] === value
 						);
 
 						await this.deleteTraverse(