+++ packages/amplify-ui-storybook/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.12](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@1.0.11...@aws-amplify/ui-storybook@1.0.12) (2021-05-14)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
+## [1.0.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@1.0.10...@aws-amplify/ui-storybook@1.0.11) (2021-05-11)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
+## [1.0.10](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@1.0.9...@aws-amplify/ui-storybook@1.0.10) (2021-05-06)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
+## [1.0.9](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@1.0.8...@aws-amplify/ui-storybook@1.0.9) (2021-04-15)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
 ## [1.0.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@1.0.7...@aws-amplify/ui-storybook@1.0.8) (2021-03-25)
 
 **Note:** Version bump only for package @aws-amplify/ui-storybook