+++ packages/amplify-ui-vue/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.9](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-vue@1.0.8...@aws-amplify/ui-vue@1.0.9) (2021-05-14)
+
+**Note:** Version bump only for package @aws-amplify/ui-vue
+
+
+
+
+
+## [1.0.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-vue@1.0.7...@aws-amplify/ui-vue@1.0.8) (2021-05-06)
+
+**Note:** Version bump only for package @aws-amplify/ui-vue
+
+
+
+
+
+## [1.0.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-vue@1.0.6...@aws-amplify/ui-vue@1.0.7) (2021-04-15)
+
+**Note:** Version bump only for package @aws-amplify/ui-vue
+
+
+
+
+
 ## [1.0.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-vue@1.0.5...@aws-amplify/ui-vue@1.0.6) (2021-03-25)
 
 **Note:** Version bump only for package @aws-amplify/ui-vue