+++ packages/api/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.3.1...@aws-amplify/api@3.3.2) (2021-05-14)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [3.3.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.3.0...@aws-amplify/api@3.3.1) (2021-05-11)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+# [3.3.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.2.30...@aws-amplify/api@3.3.0) (2021-05-06)
+
+
+### Features
+
+* **@aws-amplify/datastore:** DataStore - Multi-Auth ([#8008](https://github.com/aws-amplify/amplify-js/issues/8008)) ([dedd564](https://github.com/aws-amplify/amplify-js/commit/dedd5641dfcfce209433088fe9570874cd810997))
+
+
+
+
+
+## [3.2.30](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.2.29...@aws-amplify/api@3.2.30) (2021-04-15)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
 ## [3.2.29](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.2.28...@aws-amplify/api@3.2.29) (2021-03-25)
 
 **Note:** Version bump only for package @aws-amplify/api